Rename dims_view to qtdims
